Abstracted away the creation of SDL_Surfaces for images
Since there are numerous image types, created a factory for creating an SDL_surface regardless of type of the image (currently only PNG and BMP supported).

As a test a black king's piece is drawn on the screen. However a nullptr error occurs on a unique_ptr.

More debugging needed!