Starting to shape the game in MVC pattern
The MVC pattern is one of the most convenient design pattern to use in game development.

In this commit the board and its view is added. A game controller is added to control the game and direct the events.

This is more of a skeleton start, rather then anything new!